+++ src/configs/no-statements.ts
@@ -1,17 +1,18 @@
-import { type Linter } from "@typescript-eslint/utils/ts-eslint";
+import { type FlatConfig } from "@typescript-eslint/utils/ts-eslint";
 
-import * as noConditionalStatements from "#eslint-plugin-functional/rules/no-conditional-statements";
-import * as noExpressionStatements from "#eslint-plugin-functional/rules/no-expression-statements";
-import * as noLoopStatements from "#eslint-plugin-functional/rules/no-loop-statements";
-import * as noReturnVoid from "#eslint-plugin-functional/rules/no-return-void";
+import { rules } from "#eslint-plugin-functional/rules";
+import { ruleNameScope } from "#eslint-plugin-functional/utils/misc";
 
-const config: Linter.Config = {
-  rules: {
-    [`functional/${noConditionalStatements.name}`]: "error",
-    [`functional/${noExpressionStatements.name}`]: "error",
-    [`functional/${noLoopStatements.name}`]: "error",
-    [`functional/${noReturnVoid.name}`]: "error",
-  },
-};
-
-export default config;
+export default Object.fromEntries(
+  Object.entries(rules)
+    .filter(
+      ([, rule]) =>
+        rule.meta.deprecated !== true &&
+        rule.meta.docs.category === "No Statements" &&
+        rule.meta.docs.recommended !== false,
+    )
+    .map(([name, rule]) => [
+      `${ruleNameScope}/${name}`,
+      rule.meta.docs.recommendedSeverity,
+    ]),
+) satisfies FlatConfig.Config["rules"];