+++ src/rules/no-expression-statements.ts
@@ -14,11 +14,14 @@ import {
   shouldIgnorePattern,
   type IgnoreCodePatternOption,
 } from "#eslint-plugin-functional/options";
-import { isDirectivePrologue } from "#eslint-plugin-functional/utils/misc";
+import {
+  ruleNameScope,
+  isDirectivePrologue,
+} from "#eslint-plugin-functional/utils/misc";
 import {
   createRule,
   getTypeOfNode,
-  type NamedCreateRuleMetaWithCategory,
+  type NamedCreateRuleCustomMeta,
   type RuleResult,
 } from "#eslint-plugin-functional/utils/rule";
 import {
@@ -30,6 +33,11 @@ import {
  * The name of this rule.
  */
 export const name = "no-expression-statements" as const;
+
+/**
+ * The full name of this rule.
+ */
+export const fullName = `${ruleNameScope}/${name}` as const;
 
 /**
  * The options this rule can take.
@@ -79,11 +87,13 @@ const errorMessages = {
 /**
  * The meta data for this rule.
  */
-const meta: NamedCreateRuleMetaWithCategory<keyof typeof errorMessages> = {
+const meta: NamedCreateRuleCustomMeta<keyof typeof errorMessages> = {
   type: "suggestion",
   docs: {
     category: "No Statements",
     description: "Disallow expression statements.",
+    recommended: "recommended",
+    recommendedSeverity: "error",
     requiresTypeChecking: true,
   },
   messages: errorMessages,