+++ src/rules/no-mixed-types.ts
@@ -2,9 +2,10 @@ import { AST_NODE_TYPES, type TSESTree } from "@typescript-eslint/utils";
 import { type JSONSchema4 } from "@typescript-eslint/utils/json-schema";
 import { type RuleContext } from "@typescript-eslint/utils/ts-eslint";
 
+import { ruleNameScope } from "#eslint-plugin-functional/utils/misc";
 import {
   createRuleUsingFunction,
-  type NamedCreateRuleMetaWithCategory,
+  type NamedCreateRuleCustomMeta,
   type RuleResult,
 } from "#eslint-plugin-functional/utils/rule";
 import {
@@ -18,6 +19,11 @@ import {
  * The name of this rule.
  */
 export const name = "no-mixed-types" as const;
+
+/**
+ * The full name of this rule.
+ */
+export const fullName = `${ruleNameScope}/${name}` as const;
 
 /**
  * The options this rule can take.
@@ -67,13 +73,15 @@ const errorMessages = {
 /**
  * The meta data for this rule.
  */
-const meta: NamedCreateRuleMetaWithCategory<keyof typeof errorMessages> = {
+const meta: NamedCreateRuleCustomMeta<keyof typeof errorMessages> = {
   type: "suggestion",
   docs: {
     category: "No Other Paradigms",
     description:
       "Restrict types so that only members of the same kind are allowed in them.",
     requiresTypeChecking: true,
+    recommended: "recommended",
+    recommendedSeverity: "error",
   },
   messages: errorMessages,
   schema,