+++ src/utils/constants.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/naming-convention -- This is a special var.
+export const __VERSION__ = "0.0.0-development";