+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 # Changelog
 All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
+# [6.2.0-next.3](https://github.com/eslint-functional/eslint-plugin-functional/compare/v6.2.0-next.2...v6.2.0-next.3) (2024-03-14)
+
+
+### Features
+
+* move flat configs to new "flat" subpackage ([1d3533f](https://github.com/eslint-functional/eslint-plugin-functional/commit/1d3533fca47673ca1dc9e50dbe5a00ad8adb5362))
+
+# [6.2.0-next.2](https://github.com/eslint-functional/eslint-plugin-functional/compare/v6.2.0-next.1...v6.2.0-next.2) (2024-03-14)
+
+
+### Bug Fixes
+
+* build types ([054a862](https://github.com/eslint-functional/eslint-plugin-functional/commit/054a86299e5b6c97a78f72a7711938273d56ef82))
+* improve types ([2340489](https://github.com/eslint-functional/eslint-plugin-functional/commit/23404897399caa4f70ed0d5eb24b9ab300b65600))
+
+# [6.2.0-next.1](https://github.com/eslint-functional/eslint-plugin-functional/compare/v6.1.1...v6.2.0-next.1) (2024-03-13)
+
+
+### Features
+
+* add support for flat configs ([901d52a](https://github.com/eslint-functional/eslint-plugin-functional/commit/901d52a889c4ecf5fa78474b1f7361edb6b9e60f))
+
 ## [6.1.1](https://github.com/eslint-functional/eslint-plugin-functional/compare/v6.1.0...v6.1.1) (2024-03-11)